Implement `#left_joining_to`
Use of `#outer` -- other than at the tail-end of an association
chain -- is somewhat ambiguous to how a `LEFT OUTER JOIN` can be
usefully accomplished with `INNER JOIN`s occurring afterward.  This
new syntax allows easily building a sub-`SELECT` to which the lefthand
side is `LEFT OUTER JOIN`ed.